containers folder restructure
