+++ client/src/containers/Repos/Repos.css
@@ -1,0 +1,9 @@
+.empty-state {
+  font-size: 2rem;
+  margin: 2rem;
+  text-align: center;
+}
+
+.container {
+  margin-top: 4rem;
+}
