--- client/src/containers/Searches.css
@@ -1,4 +1,0 @@
-.loader {
-  margin: 3rem;
-  font-size: 2rem;
-}
