+++ client/src/containers/Searches/Searches.css
@@ -1,0 +1,9 @@
+.loader {
+  margin: 3rem;
+  font-size: 2rem;
+  text-align: center;
+}
+
+.container {
+  margin-top: 4rem;
+}
