+++ client/src/containers/Searches/Searches.js
@@ -1,17 +1,12 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import SearchForm from "../components/SearchForm/SearchForm";
+import SearchForm from "../../components/SearchForm/SearchForm";
 import "./Searches.css";
 
 const Searches = props => {
   const isSearching = useSelector(state => state.repos.searching);
   return (
-    <div
-      className="container"
-      style={{
-        marginTop: "4rem"
-      }}
-    >
+    <div className="container">
       {isSearching ? (
         <p className="loader">Searching Repos...</p>
       ) : (
